service/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
urllib3==2.2.3
python-dotenv==1.0.1
linked-claims-extractor>=0.1.6
requests>=2.32.3
PyJWT>=2.8.0
